week-6-geoshield-api/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
asgiref==3.11.0
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0.1
django-filter==25.2
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-yasg==1.21.14
frozenlist==1.8.0
geoip2==5.2.0
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.7
graphql-relay==3.2.0
idna==3.11
inflection==0.5.1
maxminddb==3.0.0
multidict==6.7.1
packaging==26.0
promise==2.3
propcache==0.4.1
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
redis==7.1.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
stripe==14.3.0
text-unidecode==1.3
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
yarl==1.22.0
